refactored form into sub components
The form component was a bit large so refactored into sub components to make the code cleaner and more readable
